Add `isHighlightingDataQualityIssues` prop to `UnderpassMap` definition
The prop serves as a toggle to control the display of data quality issues on the map. When this flag is disabled, both the `waysFill` and `waysLine` will be rendered using the `theme.colors.secondary` color.

Furthermore, when the `isHighlightingDataQualityIssues` flag is set to `false`, the status information will no longer be shown on the `Popup` component.